Title: 약수 출력하기, Time: 120ms, Memory: 8MB, Status: Passed - CodeTree

+++ 241007/약수 출력하기/print-divisor.java
@@ -18,7 +18,7 @@ public class Main {
         }
         if (sq == (int) sq) {
             System.out.print(" ");
-            System.out.print(sq);
+            System.out.print((int)sq);
         }
         while (!s.isEmpty()) {
             System.out.print(" ");
